Added divmod operation with BigInt with primitive ints.

+++ src/number/big_int/ops/unsigned_division.rs
@@ -117,12 +117,12 @@ impl BigInt {
 
 	pub(crate) fn unsigned_divmod_by_small_int(&self, rhs: u32) -> (BigInt, u32) {
 		let mut quotient = Vec::with_capacity(self.digits.len());
-		let mut remainder = 0u32;
+		let mut remainder = 0u64;
 		let rhs = rhs as u64;
 		for byte in self.digits.iter().rev() {
-			let current = ((remainder as u64) << 32) + *byte as u64; // Combine carry and byte
+			let current = (remainder << 32) + *byte as u64; // Combine carry and byte
 			quotient.push((current / rhs) as u32);
-			remainder = (current % rhs) as u32; // New carry is the remainder
+			remainder = current % rhs; // New carry is the remainder
 		}
 
 		(
@@ -132,7 +132,7 @@ impl BigInt {
 					.rev()
 					.collect::<Vec<u32>>()
 			),
-			remainder,
+			remainder as u32,
 		)
 	}
 }
